refactor: address review feedback for test_utils fold
- Consolidate `assert_result_error_with_message`: delete the kernel-internal
  copy in `crate::utils::test_utils`, propagate `#[track_caller]` and the
  better panic message to `crate::test_utils`, and switch all in-kernel
  callers to use it.
- Drop stale temporal references that described the old `test_utils` crate
  (table_builder.rs module/macro/TestTable docs, snapshot/mod.rs comment,
  Cargo.toml dev-dep comment).
- Strip redundant feature listings from consumer Cargo.tomls (acceptance,
  benchmarks, ffi) -- `test-utils` already implies
  `default-engine-rustls + internal-api + prettyprint`.
- Update CLAUDE.md, docs/user-guide/concepts/architecture.md and
  feature_flags.md to remove the deleted `test_utils` crate row and reflect
  the heavyweight `test-utils` feature.
- Tidy `kernel/src/test_utils/mod.rs` header and inline comments.
