+++ ffi/Cargo.toml
@@ -38,7 +38,7 @@ paste = "1.0"
 rand = "0.9.2"
 serde = "1.0.219"
 serde_json = "1.0.142"
-delta_kernel = { path = "../kernel", default-features = false, features = ["internal-api", "test-utils"] }
+delta_kernel = { path = "../kernel", default-features = false, features = ["test-utils"] }
 tokio = { version = "1.47" }
 trybuild = "1.0"
 tempfile = "3.20.0"
