+++ kernel/src/engine/arrow_data.rs
@@ -544,7 +544,8 @@ mod tests {
     use crate::expressions::ArrayData;
     use crate::schema::{ArrayType, ColumnName, DataType, StructField, StructType};
     use crate::table_features::TableFeature;
-    use crate::utils::test_utils::{assert_result_error_with_message, string_array_to_engine_data};
+    use crate::test_utils::assert_result_error_with_message;
+    use crate::utils::test_utils::string_array_to_engine_data;
     use crate::{DeltaResult, Engine as _, EngineData as _};
 
     #[test]
