+++ kernel/src/engine/arrow_expression/tests.rs
@@ -26,7 +26,7 @@ use crate::kernel_predicates::{
     IndirectDataSkippingPredicateEvaluator,
 };
 use crate::schema::{ArrayType, DataType as KernelDataType, MapType, StructField, StructType};
-use crate::utils::test_utils::assert_result_error_with_message;
+use crate::test_utils::assert_result_error_with_message;
 use crate::EvaluationHandlerExtension as _;
 
 #[test]
