+++ kernel/src/expressions/mod.rs
@@ -1272,7 +1272,7 @@ mod tests {
             Expression, Predicate, Scalar, Transform, UnaryExpressionOp,
         };
         use crate::schema::{ArrayType, DataType, DecimalType, MapType, StructField};
-        use crate::utils::test_utils::assert_result_error_with_message;
+        use crate::test_utils::assert_result_error_with_message;
 
         // ==================== Expression::Literal Tests ====================
