+++ kernel/src/parallel/sequential_phase.rs
@@ -206,8 +206,8 @@ impl<P: LogReplayProcessor> Iterator for SequentialPhase<P> {
 mod tests {
     use super::*;
     use crate::scan::AfterSequentialScanMetadata;
-    use crate::utils::test_utils::{assert_result_error_with_message, load_test_table};
-
+    use crate::test_utils::assert_result_error_with_message;
+    use crate::utils::test_utils::load_test_table;
     /// Core helper function to verify sequential processing with expected adds and sidecars.
     fn verify_sequential_processing(
         table_name: &str,
