+++ kernel/src/schema/variant_utils.rs
@@ -45,10 +45,8 @@ mod tests {
     use crate::actions::Protocol;
     use crate::schema::{DataType, StructField, StructType};
     use crate::table_features::TableFeature;
-    use crate::utils::test_utils::{
-        assert_result_error_with_message, assert_schema_feature_validation,
-    };
-
+    use crate::test_utils::assert_result_error_with_message;
+    use crate::utils::test_utils::assert_schema_feature_validation;
     #[test]
     fn test_is_unshredded_variant() {
         fn is_unshredded_variant(s: &DataType) -> bool {
