+++ kernel/src/table_changes/log_replay/tests.rs
@@ -17,8 +17,8 @@ use crate::schema::{DataType, SchemaRef, StructField, StructType};
 use crate::table_changes::log_replay::LogReplayScanner;
 use crate::table_configuration::TableConfiguration;
 use crate::table_features::{ColumnMappingMode, TableFeature};
-use crate::test_utils::LoggingTest;
-use crate::utils::test_utils::{assert_result_error_with_message, Action, LocalMockTable};
+use crate::test_utils::{assert_result_error_with_message, LoggingTest};
+use crate::utils::test_utils::{Action, LocalMockTable};
 use crate::{DeltaResult, Engine, Error, Predicate, Version};
 
 fn get_schema() -> SchemaRef {
